ffmpeg: add -fps_mode=passthrough
Add -fps_mode=passthrough. This is because otherwise ffmpeg may
drop frames due to timing issues. This is undesirable for testing,
as a) we do not indend to present the frames at all and b) dropping
even a single frame invalidates the CRC for the entire test.

This is because, without -fps_mode=passthrough, ffmpeg defaults to
either one of these:

cfr (1):
	   Frames will be duplicated and dropped to achieve exactly
the requested constant frame rate.

vfr (2):
	   Frames are passed through with their timestamp or dropped
so as to prevent 2 frames from having the same timestamp.

In particular, I have tested this change with the ffmpeg software
decoders, i.e. ./fluster.py run -d FFmpeg-H.264 -j1, and so on,
and it effectively improves the scores:

Without fps_mode=passthrough

vp8: 44/61
vp9: 237/303
h264: 120/135

With fps_mode=passthrough
vp8: 59/61
vp9: 240/303
h264: 128/135

It also improves the score for the v4l2 m2m stateful decoder I am
currently working on.